returns pointer to newly allocates string

+++ 0x0B-malloc_free/1-strdup.c
@@ -9,7 +9,7 @@
  */
 char *_strdup(char *str)
 {
-	int len = strlen(str) + 1;
+	size_t len = strlen(str) + 1;
 	char *mem_al;
 
 	if (str == NULL)
